+++ services/beaconblockproposer/standard/service.go
@@ -75,7 +75,7 @@ func (s *Service) Prepare(ctx context.Context, data interface{}) error {
 	if !ok {
 		return errors.New("passed invalid data structure")
 	}
-	log := log.With().Uint64("slot", uint64(duty.Slot())).Uint64("validator_index", uint64(duty.ValidatorIndex())).Logger()
+	log := log.With().Uint64("proposing_slot", uint64(duty.Slot())).Uint64("validator_index", uint64(duty.ValidatorIndex())).Logger()
 	log.Trace().Msg("Preparing")
 
 	// Fetch the validating account.
@@ -114,7 +114,7 @@ func (s *Service) Propose(ctx context.Context, data interface{}) {
 		s.monitor.BeaconBlockProposalCompleted(started, "failed")
 		return
 	}
-	log := log.With().Uint64("slot", uint64(duty.Slot())).Uint64("validator_index", uint64(duty.ValidatorIndex())).Logger()
+	log := log.With().Uint64("proposing_slot", uint64(duty.Slot())).Uint64("validator_index", uint64(duty.ValidatorIndex())).Logger()
 	log.Trace().Msg("Proposing")
 
 	var graffiti []byte